--- src/config/index.ts
@@ -1,2 +1,0 @@
-export * from './config.module';
-export * from './config.service';
