+++ src/main.ts
@@ -1,7 +1,7 @@
 import { NestFactory } from '@nestjs/core';
 import { ValidationPipe } from '@nestjs/common';
 import { DocumentBuilder, SwaggerModule } from '@nestjs/swagger';
-import { PrismaExceptionFilter } from './exception-filters';
+import { PrismaExceptionFilter } from './common';
 import { AppModule } from './app.module';
 
 async function bootstrap() {
